fixed bugs & added line indices

+++ lib/zigzagzoe.h
@@ -16,10 +16,10 @@ typedef game_t z3_t;
 
 z3_t *z3_init(bool player1_ai, bool player2_ai);
 z3_t *z3_init_h2(bool player1_ai, bool player2_ai);
-z3_t *z3_init_w(unsigned M, unsigned N, unsigned K, unsigned block_init, z3_stale_t mate,
+z3_t *z3_init_w(unsigned M, unsigned N, unsigned K, uint8_t block_init, z3_stale_t mate,
                 char tile_p1, char tile_p2, char tile_na, char tile_clog,
                 unsigned depth, bool player1_ai, bool player2_ai);
-z3_t *z3_init_h2_w(unsigned M, unsigned N, unsigned K, unsigned block_init, z3_stale_t mate,
+z3_t *z3_init_h2_w(unsigned M, unsigned N, unsigned K, uint8_t block_init, z3_stale_t mate,
                 char tile_p1, char tile_p2, char tile_na, char tile_clog,
                 unsigned depth, bool player1_ai, bool player2_ai);
 void z3_reset(z3_t *game);
